requirements.txt
pylint
pytest
sqlalchemy
psycopg2-binary==2.9.6
